+++ spacs/__init__.py
@@ -1,7 +1,8 @@
-from .client import SpacsClient, SpacsRequestError, ContentType
+from .client import SpacsClient, SpacsRequest, SpacsRequestError, ContentType
 
 __all__ = [
     "SpacsClient",
+    "SpacsRequest",
     "SpacsRequestError",
     "ContentType",
 ]